✒️Feat: 15회차 페어 프로그래밍에 앞선 준비
