+++ 파랑/15회차/assets/css/ui/modal.css
@@ -1,0 +1,134 @@
+/*
+Copyright (c) 2021 by Timothy Long (https://codepen.io/timothylong/pen/HhAer)
+
+Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+*/
+.modal-window {
+  position: fixed;
+  background-image: linear-gradient(to right, #0097c4 0, #adecff 100%);
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  z-index: 999;
+  visibility: hidden;
+  opacity: 0;
+  pointer-events: none;
+  transition: all 0.3s;
+}
+
+.modal-window:target {
+  visibility: visible;
+  opacity: 1;
+  pointer-events: auto;
+}
+
+.modal-window>div {
+  width: 400px;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  padding: 2em;
+  background: white;
+}
+
+.modal-window header {
+  font-weight: bold;
+}
+
+.modal-window h1 {
+  font-size: 150%;
+  margin: 0 0 15px;
+}
+
+.modal-close {
+  color: #aaa;
+  line-height: 50px;
+  font-size: 80%;
+  position: absolute;
+  right: 0;
+  text-align: center;
+  top: 0;
+  width: 70px;
+  text-decoration: none;
+}
+
+.modal-close:hover {
+  color: black;
+}
+
+.modal-window>div {
+  border-radius: 1rem;
+}
+
+.modal-window div:not(:last-of-type) {
+  margin-bottom: 15px;
+}
+
+.btn {
+  background-color: white;
+  padding: 1em 1.5em;
+  border-radius: 1rem;
+  text-decoration: none;
+}
+
+.btn i {
+  padding-right: 0.3em;
+}
+
+/* Input */
+
+.input-wrap {
+  width: 100%;
+  max-width: 30em;
+  margin: auto;
+  margin-top: 3em;
+}
+
+.input {
+  border-style: none;
+  border-bottom: var(--color-input) solid 3px;
+  padding-left: 0.3em;
+  font-size: 1.2em;
+  transition: all 300ms ease-in-out;
+}
+
+.input:focus {
+  outline: none;
+  border-bottom: var(--color-input-focus) solid 3px;
+}
+
+.submit {
+  font-size: 1em;
+  padding: 0.4em 1.4em;
+  border: none;
+  border-radius: 3em;
+  cursor: pointer;
+  transition: all 150ms ease-in-out;
+}
+
+.users {
+  background-color: lightgray;
+  padding: 1.5em;
+  border-radius: 30px;
+  margin: auto;
+  margin-top: 3em;
+}
+
+.user {
+  display: flex;
+  margin-bottom: 0.5em;
+  justify-content: space-between;
+  padding-left: 0.3em;
+}
+
+.user-name {
+  width: 100%;
+  border-bottom: var(--color-light-grey) solid 1px;
+  margin-right: 5px;
+}
