+++ 파랑/15회차/src/App.js
@@ -1,5 +1,6 @@
 import Header from './components/header/Header.js';
 import Main from './components/Main/Main.js';
+import Modal from './components/modal/Modal.js';
 
 export default class App {
   constructor(dom) {
@@ -8,6 +9,7 @@ export default class App {
     this.render();
     new Header(document.querySelector('header'));
     new Main(document.querySelector('main'));
+    new Modal(document.querySelector('#open-modal'));
   }
 
   render() {
