+++ 파랑/15회차/src/components/main/Main.js
@@ -1,6 +1,4 @@
 import Timer from './timer/Timer.js';
-// import Progressbar from './progressbar/Progressbar.js';
-import Member from './member/Member.js';
 
 export default class Main {
   constructor(dom, props) {
@@ -13,7 +11,6 @@ export default class Main {
   render(props = { editable: true }) {
     this.dom.innerHTML = `
       ${Timer(props)}
-      ${Member(props)}
     `;
   }
 }
