--- 파랑/15회차/src/components/main/progressbar/Progressbar.js
@@ -1,18 +1,0 @@
-const Progressbar = () => {
-  return `
-        <section class="progressbar-wrap">
-            <div class="roles">
-                <div class="driver">
-                    <span class="driver-name">파랑</span>
-                    <span class="driver-icon">🚗</span>
-                </div>
-                <div class="navigator">
-                    <span class="navigator-name">아벤</span>
-                    <span class="navigator-icon">😎</span>
-                </div>
-            </div>
-            <div class="progressbar"></div>
-        </section>
-        `;
-};
-export default Progressbar;
