+++ 파랑/15회차/src/components/main/timer/Timer.js
@@ -1,26 +1,28 @@
 const Timer = ({ editable, timer = '05:00', progress = '50' }) => {
   return `
-        <section class="timer-wrap">
-            <div class="timer" contenteditable="${editable}">${timer}</div>
-            <div class="timer-buttons">
-                <button class="button">Swap</button>
-                <button class="button">Clear</button>
-                <button class="button">Play</button>
-            </div>
-        </section>
-        <section class="progressbar-wrap">
-            <div class="roles">
-                <div class="driver" style="left:${progress}%;">
-                    <span class="driver-name">파랑</span>
-                    <span class="driver-icon">🚗</span>
+        <content>
+            <section class="timer-wrap">
+                <div class="timer" contenteditable="${editable}">${timer}</div>
+                <div class="timer-buttons">
+                    <button class="button">Swap</button>
+                    <button class="button">Clear</button>
+                    <button class="button">Play</button>
                 </div>
-                <div class="navigator">
-                    <span class="navigator-name">아벤</span>
-                    <span class="navigator-icon">😎</span>
+            </section>
+            <section class="progressbar-wrap">
+                <div class="roles">
+                    <div class="driver" style="left:${progress}%;">
+                        <span class="driver-name">파랑</span>
+                        <span class="driver-icon">🚗</span>
+                    </div>
+                    <div class="navigator">
+                        <span class="navigator-name">아벤</span>
+                        <span class="navigator-icon">😎</span>
+                    </div>
                 </div>
-            </div>
-            <div class="progressbar"></div>
-        </section>
+                <div class="progressbar"></div>
+            </section>
+        </content>
         `;
 };
