+++ 파랑/15회차/src/index.js
@@ -1,3 +1,9 @@
 import App from './App.js';
 
 const app = new App(document.querySelector('#root'));
+
+['.pie1', '.pie2', '.pie3'].forEach(selector => {
+  document.querySelector(selector).addEventListener('click', () => {
+    document.body.classList.remove('active');
+  });
+});
